Modify main function to print answers
- PR 코멘트 반영, main 함수에는 각 결과값이 터미널에 출력되도록 수정
