+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod leetcode;
+
+pub use crate::leetcode::vector;
+
